Addresses header QA
