+++ src/components/publishing/__stories__/article.story.tsx
@@ -1,7 +1,7 @@
 import { storiesOf } from "@storybook/react"
 import * as React from "react"
 import Article from "../article"
-import { FeatureArticle, StandardArticle } from "../fixtures/articles"
+import { FeatureArticle, MissingVerticalStandardArticle, StandardArticle } from "../fixtures/articles"
 import { RelatedCanvas, RelatedPanel } from "../fixtures/components"
 
 storiesOf("Publishing/Articles", module)
@@ -9,6 +9,16 @@ storiesOf("Publishing/Articles", module)
     return (
       <Article
         article={StandardArticle}
+        relatedArticlesForPanel={RelatedPanel}
+        relatedArticlesForCanvas={RelatedCanvas}
+        emailSignupUrl="#"
+      />
+    )
+  })
+  .add("Standard without Vertical", () => {
+    return (
+      <Article
+        article={MissingVerticalStandardArticle}
         relatedArticlesForPanel={RelatedPanel}
         relatedArticlesForCanvas={RelatedCanvas}
         emailSignupUrl="#"
